+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Remove `backtrace-rs` feature, as the default choice when not specified (#130)
 
+### Fixed
+- Export `UnresolvedReport` type to allow developers to get the unresolved report (#132)
+
 ## [0.9.1] - 2022-05-19
 
 ### Fixed